decouple module action helper implementation
